Attempt to fix
PVP-156-1: Documentation~/filter.yml uses sensible regexes
Checks that regex patterns are anchored and periods are escaped.

Both the beginning and end of a pattern should be anchored with respectively ^ and $, or alternatively, it should be made explicit that the pattern is unanchored using \. or .*.

Example: Instead of OnUpdate, do \.OnUpdate$. Instead of Tests$, do \.Tests$ or .*Tests$.

(Though note that the proper fix for hiding tests is usually to make the test classes internal instead of public.)

A period should be either escaped (as \.) or quantified (e.g. .*), or alternatively it should be made explicit that it's not escaped by surrounding it with a capture group using (.).

Example: Instead of Tests.Runtime, do .*Tests\.Runtime.*.

For more information, see the filter.yml docs.

#298019: com.unity.mobile.android-logcat@1.4.4, 2021.3.50f1, macos
Check failed with 2 error(s):

Documentation~/filter.yml: regex should be anchored: ^Global
Documentation~/filter.yml: regex should be anchored: ^Unity\.Android\.Logcat